combine ShareLink and Url
