+++ src/components/content-types/TopicList.tsx
@@ -6,7 +6,7 @@ import "./TextContent.css"
 import { DocCollection, DocHandle, DocumentId } from "automerge-repo"
 import ContentList, { ContentListInList } from "./ContentList"
 import { BoardDoc, BoardDocCard } from "./board"
-import { parseDocumentLink, PushpinUrl } from "../pushpin-code/ShareLink"
+import { parseDocumentLink, PushpinUrl } from "../pushpin-code/Url"
 import { TextDoc } from "./TextContent"
 import "./TopicList.css"
 import classNames from "classnames"
