+++ src/components/content-types/files/FileContent.tsx
@@ -2,7 +2,7 @@ import React, { useRef } from "react"
 import Content, { ContentProps } from "../../Content"
 import * as ContentTypes from "../../pushpin-code/ContentTypes"
 import { useDocument } from "automerge-repo-react-hooks"
-import { createDocumentLink } from "../../pushpin-code/ShareLink"
+import { createDocumentLink } from "../../pushpin-code/Url"
 import { FileDoc } from "."
 
 import "./FileContent.css"
