+++ src/components/content-types/workspace/WorkspaceInList.tsx
@@ -3,7 +3,7 @@ import Content, { ContentProps } from "../../Content"
 import { useDocument } from "automerge-repo-react-hooks"
 import Badge from "../../ui/Badge"
 import { WorkspaceDoc } from "./Workspace"
-import { createDocumentLink } from "../../pushpin-code/ShareLink"
+import { createDocumentLink } from "../../pushpin-code/Url"
 import { ContactDoc } from "../contact"
 import "./WorkspaceInList.css"
 import ListItem from "../../ui/ListItem"
