+++ src/components/content-types/workspace/omnibox/ActionListItem.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react"
 import Actions, { ActionItem } from "./Actions"
-import { PushpinUrl } from "../../../pushpin-code/ShareLink"
+import { PushpinUrl } from "../../../pushpin-code/Url"
 import ListMenuItem, { Stretch } from "../../../ui/ListMenuItem"
 
 export interface Props {
