+++ src/components/content-types/workspace/omnibox/InvitationsHook.ts
@@ -1,7 +1,7 @@
 import { DocumentId } from "automerge-repo"
 import { useDocument } from "automerge-repo-react-hooks"
 import { useDocumentIds, useDocuments } from "../../../pushpin-code/Hooks"
-import { PushpinUrl } from "../../../pushpin-code/ShareLink"
+import { PushpinUrl } from "../../../pushpin-code/Url"
 import { ContactDoc } from "../../contact"
 import { WorkspaceDoc } from "../Workspace"
