+++ src/components/ui/ContentDragHandle.tsx
@@ -3,7 +3,7 @@ import "./Heading.css"
 import mime from "mime-types"
 import * as UriList from "../pushpin-code/UriList"
 import "./ContentDragHandle.css"
-import { PushpinUrl } from "../pushpin-code/ShareLink"
+import { PushpinUrl } from "../pushpin-code/Url"
 import { DocumentId } from "automerge-repo"
 import { BinaryDataId, useBinaryDataHeader } from "../../blobstore/Blob"
